+++ src/components/nb-custom-accent-modal/custom-accent-modal.css
@@ -1,0 +1,262 @@
+@import "../../css/colors.css";
+@import "../../css/units.css";
+
+.modal-content {
+    max-width: 550px;
+    margin-top: 50px;
+}
+
+.body {
+    background: $ui-white;
+    padding: 1.5rem 2.25rem;
+    max-height: calc(100vh - 100px - 3.125rem);
+    overflow: auto;
+}
+
+.body p,
+.input,
+.text-code-input {
+    margin: 1rem 0;
+    display: block;
+}
+
+.type-selector-container {
+    display: flex;
+    justify-content: space-around;
+    margin-bottom: 1rem;
+}
+.type-selector-button {
+    width: 100%;
+    cursor: pointer;
+    border-bottom: 0.25rem solid $ui-tertiary;
+    margin: 0 1rem;
+    padding: 0.5rem 0;
+    text-align: center;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+.type-selector-button[data-active="true"] {
+    border-color: $looks-secondary;
+}
+.type-selector-button:active {
+    border-color: $looks-transparent;
+}
+
+.input {
+    width: 100%;
+    border: 1px solid $ui-black-transparent;
+    border-radius: 0.25rem;
+    padding: 0.5rem;
+    font-size: inherit;
+    height: 3rem;
+}
+
+.button-row {
+    display: flex;
+    justify-content: flex-end;
+}
+
+/* Confirmation buttons at the bottom of the modal */
+.button-row {
+    margin-top: 1rem;
+    font-weight: bolder;
+    text-align: right;
+}
+
+.button-row button {
+    border: 1px solid $ui-black-transparent;
+    border-radius: 0.25rem;
+    padding: 0.75rem 1rem;
+    background: white;
+    color: black;
+    font-weight: bold;
+    font-size: 0.85rem;
+}
+
+.button-row button.ok-button {
+    background: $looks-secondary;
+    border: $looks-secondary;
+    color: white;
+}
+
+[dir="ltr"] .button-row button + button {
+    margin-left: 0.5rem;
+}
+
+[dir="rtl"] .button-row button + button {
+    margin-right: 0.5rem;
+}
+
+label {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 0.5rem;
+}
+
+.color-option {
+    cursor: pointer;
+    height: 35px;
+    width: 35px;
+    border: 2px solid $ui-black-transparent;
+    border-radius: 50%;
+    padding: calc($space * 1);
+    transition: all 0.2s;
+}
+
+.color-option:hover {
+    border: 2px solid $looks-secondary;
+    box-shadow: 0px 0px 0px 4px $looks-transparent;
+}
+
+.color-option + .color-option {
+    margin-left: 0.5rem;
+}
+
+.color-picker {
+    appearance: none;
+    -moz-appearance: none;
+    -webkit-appearance: none;
+    cursor: pointer;
+    padding: calc($space * 1);
+    border: 2px solid $ui-black-transparent;
+    border-radius: 50%;
+    width: 35px;
+    height: 35px;
+}
+
+.color-picker::-moz-color-swatch {
+    border: none;
+    border-radius: 50%;
+    padding: 0;
+}
+
+.color-picker::-webkit-color-swatch {
+    border: none;
+    border-radius: 50%;
+    padding: 0;
+}
+
+.color-picker::-webkit-color-swatch-wrapper {
+    border: none;
+    border-radius: 50%;
+    padding: 0;
+}
+
+.color-picker:hover {
+    border: 2px solid $looks-secondary;
+    box-shadow: 0px 0px 0px 4px $looks-transparent;
+}
+
+.option-card {
+    background: $ui-white;
+    border: 2px solid $ui-black-transparent;
+    border-radius: $space;
+    padding: calc($space * 2);
+    text-align: center;
+    flex-grow: 1;
+    transition: all 0.2s;
+    flex-basis: 100px;
+}
+
+.option-title {
+    font-weight: bold;
+    margin-bottom: 8px;
+}
+
+.options-row {
+    display: flex;
+    gap: 1rem;
+}
+
+.options-row > span {
+    flex-grow: 1;
+    gap: 1rem;
+}
+
+.card-box {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+}
+
+.card {
+    background: $ui-white;
+    border: 2px solid $ui-black-transparent;
+    border-radius: $space;
+    padding: calc($space * 1.5);
+    display: flex;
+    gap: 1rem;
+}
+
+.color-preview {
+    border-radius: 50% !important;
+}
+
+.card > div > *:not(input) {
+    appearance: none;
+    width: 20px;
+    height: 20px;
+    background-color: $looks-secondary;
+    background-size: 100%;
+    border: 1px solid $ui-black-transparent;
+    border-radius: 4px;
+    outline: none;
+    transition: all 0.2s;
+    margin: auto;
+}
+
+.card > div[dir="rtl"] > * {
+    margin-left: 0;
+}
+
+
+.card > div[dir="ltr"] > * {
+    margin-right: 0;
+}
+
+.card > div > button:hover {
+    border: 1px solid $looks-secondary;
+    box-shadow: 0px 0px 0px 4px $looks-transparent;
+}
+
+.card > div {
+    display: flex;
+    gap: 0.5rem;
+}
+
+.download-option {
+    background-image: url("./icon--file-download.svg");
+    background-size: 90%;
+}
+
+.edit-option {
+    background-image: url("../menu-bar/icon--settings.svg");
+}
+
+.delete-option {
+    background-image: url("../delete-button/icon--delete.svg");
+}
+
+.name {
+    flex-grow: 1;
+}
+
+.button {
+    font-family: inherit;
+    font-size: 14px;
+    background-color: $looks-secondary;
+    color: white;
+    border: none;
+    padding: 0.5rem 1rem;
+    margin: 0;
+    border-radius: 4px;
+    cursor: pointer;
+    font-family: inherit;
+    font-weight: bold;
+}
+
+.button:active, .button:focus-within {
+    box-shadow: 0 0 0 3px $looks-transparent;
+}
